Get rid of casts to scanFetchProc in pixman-bits-image.c
Instead just declare the functions with the required type and do any
type conversions in the function itself.

+++ pixman/pixman-private.h
@@ -212,8 +212,8 @@ _pixman_image_get_scanline_64_generic  (pixman_image_t *pict,
 					int             x,
 					int             y,
 					int             width,
-					uint64_t       *buffer,
-					uint64_t       *mask,
+					uint32_t       *buffer,
+					uint32_t       *mask,
 					uint32_t        maskBits);
 
 source_pict_class_t
